Create Change detection using Random forest Classifier on Google Earth Engine
This code is in Javascript, which is the major programming language for Google Earth Engine. It involves the use of the Random Forest algorithm and Landsat 8 to first create a land cover classification map and then do a change detection to monitor land cover change in the Goaso area of Ghana.