Add files via upload
In this exercise, we demonstrate how to produce maps of informal settlements out of very high resolution satellite images using fully convolutional networks. The dataset is subset of the image tiles used in [1]. The tiles came from a pan-sharpened Quickbird satellite of the city of Dar es Salaam image acquired in 2007. It is composed of four multispectral bands having a spatial resolution of 60 cm. The task is formulated as a binary classification problem distinguishing "informal settlements" from "others" class. Unlike the previous exercise in which we assign a single label for an image patch, here we classify each pixel according to the two aformentioned classes.